backend/requirements.txt
aerich
fastapi
uvicorn
aiohttp
aiosqlite
cachetools
coloredlogs
toml
tortoise-orm
requests
gitpython
nest-asyncio
python-magic